Fix compiler warnings on OSX.

+++ APElement.h
@@ -18,9 +18,7 @@
  along with APXML.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#import <UIKit/UIKit.h>
 @class APAttribute;
-
 
 @interface APElement : NSObject {
 	NSString *name;
@@ -42,8 +40,8 @@
 - (void)addAttributes:(NSDictionary*)someAttributes;
 - (void)addChild:(APElement*)anElement;
 - (void)appendValue:(NSString*)aValue;
-- (int)attributeCount;
-- (int)childCount;
+- (NSUInteger)attributeCount;
+- (NSUInteger)childCount;
 - (NSArray*)childElements;
 - (NSMutableArray*)childElements:(NSString*)aName;
 - (APElement*)firstChildElement;
